feat: Refactor questionnaire and presets handling
- Introduced a new `presets.ts` file to manage preset definitions and logic.
- Added functions to load and expand presets, including user choice resolution.
- Updated `composer.ts` to utilize `CompositionInput` for answers, enhancing type safety.
- Created a new `questionnaire.ts` file to handle the interactive questionnaire flow.
- Implemented functions for each section of the questionnaire, including stack, overlays, and output configuration.
- Added manifest loading and validation in `manifest-migrations.ts`, ensuring proper error handling.
- Enhanced project configuration with new functions to find and build answers from manifests.
- Introduced `resolveRepoPath` utility to handle path resolution for both source and compiled outputs.
- Updated types in `types.ts` to include `CompositionInput`, ensuring consistency across the application.
