+++ CLAUDE.md
@@ -1,0 +1,34 @@
+# AGENTS.md
+
+## Commands
+
+- `npm run lint` — TypeScript type-check + Prettier (CI-enforced; run before committing)
+- `npm run lint:fix` — auto-fix Prettier formatting issues (run after adding/changing files)
+- `npm test` — Vitest unit tests
+- `npm run build` — compile TypeScript to `dist/`
+- `npm run docs:generate` — regenerate `docs/overlays.md` (run after adding/changing overlays)
+- `npm run init` — run CLI from source (tsx); `npm run init:build` from compiled output
+
+## Do not edit
+
+- `dist/` — compiled output, regenerated by `npm run build`
+- `.devcontainer/` in any generated project — edit `superposition.json` + overlay files, then run `regen`
+- `docs/overlays.md` — generated; edit overlay files, then run `npm run docs:generate`
+- `overlays/index.yml` — legacy file, superseded by per-overlay `overlay.yml`; do not add entries here
+
+## Non-obvious rules
+
+- **ESM imports**: always use `.js` extension in imports, even when importing a `.ts` file (e.g. `import { foo } from './bar.js'`)
+- **Path resolution**: any new `__dirname`-based path must use a candidate array covering both source (`scripts/`) and compiled (`dist/scripts/`) locations — see existing `OVERLAYS_DIR_CANDIDATES` pattern in `scripts/init.ts`
+- **Adding an overlay**: create `overlays/<id>/overlay.yml` (id must match directory name exactly) + `devcontainer.patch.json` + `README.md`; if it introduces a new category, also update `tool/schema/types.ts` and the loader/composer in `tool/questionnaire/composer.ts` together in the same change
+- **Conflicts are bidirectional**: if overlay A lists B in `conflicts`, B must also list A
+- **Docker Compose networks**: never use `external: true`; always declare the network inline with `name: devnet`
+- **Spec-first**: before writing implementation code for a new feature, commit a spec under `docs/specs/`
+- **Changelog required**: add an appropriate `CHANGELOG.md` entry for user-visible changes; anything introduced **for the first time in the current release** must appear as a single consolidated entry under `Added` only — do not also list it under `Changed` or `Fixed`; if iterative work on an unreleased item is done, fold all details into the single `Added` bullet rather than adding separate `Changed` or `Fixed` bullets for the same item
+- **Docs required**: run `npm run docs:generate` after adding or changing overlays, then commit the updated `docs/overlays.md`
+- **Lint:fix before committing**: run `npm run lint:fix` to auto-fix formatting issues, then verify with `npm run lint`
+
+**Key Technologies:**
+
+- **Runtime**: Node.js 20+ with TypeScript 5.3.3 (compiled to ESM modules in `dist/`)
+- **CLI Framework**: Commander for argument parsing, Inquirer for interactive prompts
