+++ tool/questionnaire/composer.ts
@@ -4,6 +4,7 @@ import { fileURLToPath } from 'url';
 import chalk from 'chalk';
 import * as yaml from 'js-yaml';
 import type {
+    CompositionInput,
     QuestionnaireAnswers,
     DevContainer,
     CloudTool,
@@ -58,26 +59,14 @@ import {
     overlaysToServices,
     portsToPortInfo,
 } from '../utils/summary.js';
+import { resolveRepoPath } from '../utils/paths.js';
 
 // Get __dirname equivalent in ESM
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-// Resolve REPO_ROOT that works in both source and compiled output
-// When running from TypeScript sources (e.g. ts-node), __dirname is "<root>/tool/questionnaire"
-// When running from compiled JS in "dist/tool/questionnaire", __dirname is "<root>/dist/tool/questionnaire"
-const REPO_ROOT_CANDIDATES = [
-    path.join(__dirname, '..', '..'), // From source: tool/questionnaire -> root
-    path.join(__dirname, '..', '..', '..'), // From dist: dist/tool/questionnaire -> root
-];
-const REPO_ROOT =
-    REPO_ROOT_CANDIDATES.find(
-        (candidate) =>
-            fs.existsSync(path.join(candidate, 'templates')) &&
-            fs.existsSync(path.join(candidate, 'overlays'))
-    ) ?? REPO_ROOT_CANDIDATES[0];
-
-const TEMPLATES_DIR = path.join(REPO_ROOT, 'templates');
+const TEMPLATES_DIR = resolveRepoPath('templates', path.join(__dirname, '..', '..'));
+const REPO_ROOT = path.dirname(TEMPLATES_DIR);
 
 // ─── JetBrains support ────────────────────────────────────────────────────
 
@@ -548,7 +537,7 @@ function resolveDependencies(
  * This shared logic is used by both generateManifestOnly and composeDevContainer.
  */
 function prepareOverlaysForGeneration(
-    answers: QuestionnaireAnswers,
+    answers: CompositionInput,
     overlaysDir?: string
 ): {
     overlays: string[];
@@ -666,7 +655,7 @@ function prepareOverlaysForGeneration(
  */
 function generateManifest(
     outputPath: string,
-    answers: QuestionnaireAnswers,
+    answers: CompositionInput,
     overlays: string[],
     autoResolved: { added: string[]; reason: string },
     containerName?: string,
@@ -2049,7 +2038,7 @@ function copyCustomFiles(
  * Used for team collaboration workflow where manifest is committed but .devcontainer is gitignored
  */
 export async function generateManifestOnly(
-    answers: QuestionnaireAnswers,
+    answers: CompositionInput,
     overlaysDir?: string,
     options: { isRegen?: boolean } = {}
 ): Promise<GenerationSummary> {
@@ -2116,7 +2105,7 @@ export async function generateManifestOnly(
  * Main composition logic
  */
 export async function composeDevContainer(
-    answers: QuestionnaireAnswers,
+    answers: CompositionInput,
     overlaysDir?: string,
     options: { isRegen?: boolean } = {}
 ): Promise<GenerationSummary> {
